Conflicts:
	modules/enterprise/gui/coregui/src/main/resources/org/rhq/enterprise/gui/coregui/client/Messages_ja.properties
	modules/enterprise/gui/coregui/src/main/resources/org/rhq/enterprise/gui/coregui/client/Messages_zh.properties

+++ etc/scripts/install-rhq-server.sh
@@ -1,5 +1,26 @@
 #!/bin/sh
 
+#
+# This script silently installs an RHQ Server under a specified parent 
+# directory. The script will make sure the name of the RHQ Server install 
+# directory is "rhq-server", so as to keep its location consistent across 
+# upgrades and make it easier to manage via scripts. 
+# 
+# NOTE: If you are upgrading or replacing an existing set of RHQ Servers, you 
+# must make sure all of those Servers are stopped before running this script. 
+#
+# The script is configured via the following environment variables. For 
+# convenience, these variables can be defined in a file named 
+# install-rhq-server-env.sh located in the same directory as this script.
+#
+# RHQ_SERVER_DIST_URL - the URL or path of the RHQ Server distribution zipfile
+# RHQ_SERVER_PROPERTIES_URL - the URL or path of the rhq-server.properties file
+#                             containing the configuration the Server should use
+# RHQ_SERVER_INSTALL_PARENT_DIR - the path of the directory under which the 
+#                                 Server install directory (rhq-server) will be
+#                                 created (e.g. $HOME/Applications)
+#
+
 # Define functions.
 abort()
 {
@@ -12,28 +33,42 @@ abort()
 
 usage() 
 {   
-   abort "$@" "Usage:   $EXE RHQ_SERVER_DIST_URL RHQ_SERVER_PROPERTIES_URL RHQ_SERVER_INSTALL_PARENT_DIR" "Example: $EXE http://sourceforge.net/projects/rhq/files/rhq/rhq-4.0.1/rhq-server-4.0.1.zip/download rhq-server.properties /home/bob/Applications"
+   abort "$@" "Usage: $EXE"
 }
 
 # Process command line args.
 EXE=`basename $0`
-if [ "$#" -ne 3 ]; then
-   usage "Invalid number of arguments."
+if [ "$#" -ne 0 ]; then
+   usage "This script does not take any arguments. It is configured via environment variables."
 fi  
-RHQ_SERVER_DIST_URL="$1"
-RHQ_SERVER_PROPERTIES_URL="$2"
-RHQ_SERVER_INSTALL_PARENT_DIR="$3"
 
-echo "If you are upgrading or replacing existing RHQ Servers, make sure all of those Servers are stopped, then hit Enter to continue."
+if [ -z "$RHQ_SERVER_DIST_URL" ]; then
+   abort "RHQ_SERVER_DIST_URL environment variable is not defined."
+fi
+
+if [ -z "$RHQ_SERVER_PROPERTIES_URL" ]; then
+   abort "RHQ_SERVER_PROPERTIES_URL environment variable is not defined."
+fi
+
+if [ -z "$RHQ_SERVER_INSTALL_PARENT_DIR" ]; then
+   abort "RHQ_SERVER_INSTALL_PARENT_DIR environment variable is not defined."
+fi
+
+echo "*** NOTE *** If you are upgrading or replacing existing RHQ Servers, make sure *all* of those Servers are stopped, then hit Enter to continue."
 read
 
 # Download the Server zipfile.
-mkdir -p "$HOME/Downloads/tmp"
-cd "$HOME/Downloads/tmp"
-rm -rf *
-wget --content-disposition --no-check-certificate --timestamping "$RHQ_SERVER_DOWNLOAD_URL"
-RHQ_SERVER_DIST_FILE_NAME=`echo *`
-mv $RHQ_SERVER_DIST_FILE_NAME ..
+if [ -f "$RHQ_SERVER_DIST_URL" ]; then
+   RHQ_SERVER_DIST_FILE="$RHQ_SERVER_DIST_URL"
+else
+   mkdir -p "$HOME/Downloads/tmp"
+   cd "$HOME/Downloads/tmp"
+   rm -rf *
+   wget --content-disposition --no-check-certificate --timestamping "$RHQ_SERVER_DIST_URL"
+   RHQ_SERVER_DIST_FILE_NAME=`echo *`
+   mv $RHQ_SERVER_DIST_FILE_NAME ..
+   RHQ_SERVER_DIST_FILE="$HOME/Downloads/$RHQ_SERVER_DIST_FILE_NAME"
+fi
 
 # Download the Server properties file.
 if [ -f "$RHQ_SERVER_PROPERTIES_URL" ]; then
@@ -44,21 +79,25 @@ else
    rm -rf *
    wget --content-disposition --no-check-certificate --timestamping "$RHQ_SERVER_PROPERTIES_URL"
    RHQ_SERVER_PROPERTIES_FILE_NAME=`echo *`
-   RHQ_SERVER_PROPERTIES_FILE="/tmp/$RHQ_SERVER_PROPERTIES_FILE_NAME"
+   RHQ_SERVER_PROPERTIES_FILE="$HOME/Downloads/tmp/$RHQ_SERVER_PROPERTIES_FILE_NAME"
 fi
 
 # Unzip the Server zipfile.
 mkdir "$RHQ_SERVER_INSTALL_PARENT_DIR/tmp"
 cd "$RHQ_SERVER_INSTALL_PARENT_DIR/tmp"
-unzip -q "$HOME/Downloads/$RHQ_SERVER_DIST_FILE_NAME"
+unzip -q "$RHQ_SERVER_DIST_FILE"
 RHQ_SERVER_HOME_BASE_NAME=`echo *`
-RHQ_SERVER_HOME="$RHQ_SERVER_INSTALL_PARENT_DIR/$RHQ_SERVER_HOME_BASE_NAME"
+
+RHQ_SERVER_HOME="$RHQ_SERVER_INSTALL_PARENT_DIR/rhq-server"
 if [ -f "$RHQ_SERVER_HOME" ]; then
    # Backup existing Server installation.
    rm -rf "$RHQ_SERVER_HOME.bak"
    mv "$RHQ_SERVER_HOME" "$RHQ_SERVER_HOME.bak"
 fi
-mv $RHQ_SERVER_HOME_BASE_NAME ..
+# In the dist zipfile, the name of the RHQ install dir includes the version
+# (e.g. "rhq-server-4.1.0-SNAPSHOT"). Rename it to "rhq-server" to keep its 
+# location consistent across upgrades and make it easier to manage via scripts.
+mv "$RHQ_SERVER_HOME_BASE_NAME" "$RHQ_SERVER_HOME"
 
 # Install the Server properties file.
 mv -f "$RHQ_SERVER_PROPERTIES_FILE" "$RHQ_SERVER_HOME/bin/rhq-server.properties"
